Be more verbose about errors; print numeric code.
I was getting POOL_IN_USE, which is -200530, but it was showing up as
PROTOCOL_ERROR.  So, I also started using ServerException, which seems
to have been previously unused.

Eventually, I'd like to get a more human-readable version of the error
code printed, but just printing the number is a step up from not
knowing at all.