bug 9738: force source file encoding to UTF-8 to prevent test failure
So, Jao had the clever idea to include non-ASCII characters directly
in his source files, rather than using escape sequences, which would
be more portable.

Apparently, by default the javac compiler just uses whatever character
encoding it feels like.  On Linux, it feels like UTF-8, which is what
Jao was also feeling like, so the tests pass on Linux.  On the Mac,
though, javac feels like some other encoding (I don't know what, but
it doesn't really matter), so the string literal is not what Jao meant
it to be, and the test fails.

I've fixed this stupidity (of both Jao and the authors of javac) by
forcing the encoding to be UTF-8, regardless of platform.