bug 10598: don't discriminate against a server because of what it supports
The whole point of feature negotiation is that the server is probably
going to support more features than you want to use.  If there's a
feature you don't want to use, then just don't use it.  Simple as
that.  If you refuse to connect to a server just because it supports a
feature that you don't want to use, then you've broken the entire
concept of feature negotiation.